requirements_docker.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dnspython==2.4.2
emoji==2.8.0
fastjsonschema==2.16.2
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.26.0
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
plotly==5.17.0
protobuf==4.24.3
pyarrow==13.0.0
pydeck==0.8.0
Pygments==2.16.1
pymongo==4.5.0
Pympler==1.0.1
pyparsing==3.1.1
pyrsistent==0.19.3
pytest==7.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-lsp-jsonrpc==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.29.0
rich==13.5.3
rpds-py==0.10.3
safetensors==0.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.26.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
transformers==4.33.2
twitchAPI==3.11.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
ujson==5.4.0
urllib3==1.26.16
validators==0.22.0
whisper==1.1.10
yarl==1.9.2
zipp==3.17.0
